got everything setup to work with the database
