+++ app/src/main/java/com/ls/localsky/ui/app/LocalSkyApp.kt
@@ -37,7 +37,6 @@ fun LocalSkyApp(
                     databaseLS = database,
                     weatherViewModel = weatherViewModel,
                     cache = cache,
-                    innerPadding = innerPadding
                 )
             }
         }
